% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{create_universe}
\alias{create_universe}
\title{function to either bring a file with universe or read a file with universe}
\usage{
create_universe()
}
\description{
function to either bring a file with universe or read a file with universe
}
