% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{deduplicate_rows}
\alias{deduplicate_rows}
\title{Remove duplicated rows.}
\usage{
deduplicate_rows(dat)
}
\arguments{
\item{dat}{Data frame object with two columns: \strong{id} and set
labels \strong{set_label}.}
}
\value{
Data frame.
}
\description{
\code{deduplicate_rows} function removes duplicated rows prior to mapping
to ENTREZ gene identifiers.
}
