% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_pvalues.R
\name{filter_pvalues}
\alias{filter_pvalues}
\title{Filter GO enrichment results using a p-value cutoff}
\usage{
filter_pvalues(dat, which_pvalue = "adjusted", cutoff)
}
\arguments{
\item{dat}{Data frame with the results of GO enrichment analysis.}

\item{which_pvalue}{A character string specifying p-value for filtering.
Must of one of two: "raw" or "adjusted"}

\item{cutoff}{A numeric value specifying the cutoff. All rows with p-values
that are less or equal to the cutoff will be retained.}
}
\value{
Data frame.
}
\description{
\code{filter_pvalues} takes a data frame with the results of GO enrichment
analysis and performs row filtering based on provided \code{cutoff} value.
The data can be filtered either using the raw (\strong{pvalue} column) or the
adjusted (default) p-values (\strong{p_bonferroni} column).
}
\seealso{
\code{get_ontology}.
}
