% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{read_file}
\alias{read_file}
\title{Read data with genes and gene set labels to a data frame}
\usage{
read_file(path, col_names, delim)
}
\arguments{
\item{path}{Character string with the file path.}

\item{col_names}{A logical value indicating whether the file contains column
names in the first line.}

\item{delim}{Character string indicating the field separator character.}
}
\value{
Data frame with two columns, where the first column is \strong{id}
(gene identifiers) and the second column is \strong{set_label}
(gene set labels).
}
\description{
\code{read_file} reads a file given a path, column names and a field
delimiter. For the file to be processed correctly ensure that the gene
identifiers are in the first columns and the set labels are in the second
column in the file.
}
