% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not_exported.R
\name{remove_errors}
\alias{remove_errors}
\title{Remove sets which generated errors or warnings by \code{get_all_ontologies}.}
\usage{
remove_errors(dat)
}
\arguments{
\item{dat}{Data frame, output of the parallelized run.}
}
\value{
Data frame.
}
\description{
\code{remove_errors} removes rows in the output of \code{get_all_ontologies} for sets
which generated errors or warnings. Also provides a helpful message listing
the labels of the removed sets. This is an internal function.
}
\seealso{
\code{\link{run_parallel_go}}.
}
