% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_GO.R
\name{run_GO}
\alias{run_GO}
\title{Main function for running GO analysis}
\usage{
run_GO(path, col_names, delim, min_set_size, species, id)
}
\arguments{
\item{path}{The name of the fle which the data are to be read from}

\item{col_names}{a logical value indicating whether the file contains the names
of the variables as its first line}

\item{delim}{The file field separator character.}

\item{min_set_size}{Positive integer. If provided will be used to remove
all gene sets where then number of genes is fewer than \code{min.set.size}.}

\item{species}{A character string specifying the species, must be one of two
"human" (default) or "mouse". For mapping to Entrez Gene identifiers we use
'org.Hs.eg.db' (human) and 'org.Mm.eg.db' (mouse) Bioconductor annotation
packages.}

\item{id}{A character string specifying the type of gene identifier in the
data, must of be one of two "hugo" (default) or "ensembl".}
}
\description{
Main function for running GO analysis
}
