% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ontologies.R
\name{run_parallel_go}
\alias{run_parallel_go}
\title{Run GO analysis on parallel}
\usage{
run_parallel_go(dat, background, species = c("human", "mouse"), cores,
  ontology)
}
\arguments{
\item{annotation}{string providing the name of the annotation package to use,
Can we use anything else but the 'org.Hs.eg.db' forhuman and 'org.Mm.eg.db' for
mouse? I haven't investigated
@param ontology vector of ontologies for which to run the analysis. If not provided
then the analysis will be run for all ontologies: CC, BP and MF}
}
\description{
Run GO analysis on parallel
}
