% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not_exported.R
\name{verify_input}
\alias{verify_input}
\title{Supporting function for verifying function parameters}
\usage{
verify_input(input_name, input_choices, input_default)
}
\description{
Not to be used by the user. Not exported
}
