% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_data}
\alias{clean_data}
\title{Basic data cleaning before gene mapping}
\usage{
clean_data(dat, min_set_size)
}
\arguments{
\item{dat}{Data frame object with two columns: genes (\code{id}) and set labels (\code{set.name}),
output of the function \code{\link{read_data}} call.}

\item{min_set_size}{Positive integer. If provided will be used to remove
all gene sets where then number of genes is fewer than \code{min.set.size}.}
}
\value{
Data frame object
}
\description{
\code{clean_data} function removes duplicated rows, sets that have fewer genes
than \code{min_set_size}.
}
