% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ontologies.R
\name{get_all_ontologies}
\alias{get_all_ontologies}
\title{Perform GO enrichment given the list of ontologies.}
\usage{
get_all_ontologies(dat, universe, species = c("human", "mouse"),
  ontologies, set_label)
}
\arguments{
\item{dat}{Vector with the list of ENTREZ gene identifiers.}

\item{universe}{Vector of unique ENTREZ gene ids to be used as a background.}

\item{species}{Character string specifying the species, must be one of two
"human" (default) or "mouse". This will define which Bioconductor package
is used as a source of GO annotations. For human we use "org.Hs.eg.db", and
for mouse we use "org.Mm.eg.db".}

\item{ontologies}{Character string specifying one to three ontologies:
"MF", "BP" or "CC".}

\item{set_label}{Character string specifying the label characterizing the
set of genes being analyzed.}
}
\value{
Data frame with the results of GO enrichment and a column populated
with the provided "set_label".
}
\description{
Returns GO enrichment results for all specified GO  without p value filtering,
but with Bonferroni adjustment.
}
