% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ontologies.R
\name{get_all_ontologies}
\alias{get_all_ontologies}
\title{Run gene ontology given a list of ENTREZ GENE IDs and backgroun}
\usage{
get_all_ontologies(dat, universe, species = c("human", "mouse"),
  ontology)
}
\arguments{
\item{dat}{data frame with two columns: gene_id and module. Module provides information
about group assignment, gene_id is the list of ENTREZ GENE IDs. Module is useful
for parallel computation}

\item{universe}{vector of unique ENTREZ gene ids to use as a background}

\item{species}{a character string specifying the species, must be one of two "human"
(default) or "mouse". For GO analysis we use 'org.Hs.eg.db' for
human and 'org.Mm.eg.db' for mouse.}
}
\description{
Returns BP, MF and CC results without p value filtering, but with Bonferroni
adjustment.
}
