% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_genes.R
\name{map_genes}
\alias{map_genes}
\title{Map gene identifiers to ENTREZ gene}
\usage{
map_genes(dat, id = "hugo", species = "human")
}
\arguments{
\item{dat}{Data frame with two columns: \strong{id} and \strong{set_label}.}

\item{id}{A character string specifying the type of gene identifier in the
data, must of be one of two "hugo" (default) or "ensembl".}

\item{species}{A character string specifying the species, must be one of two
"human" (default) or "mouse".}
}
\value{
Data frame with two columns: \strong{entrez} and \strong{set_label}.
The column with the original identifiers is removed.
}
\description{
\code{map_genes} provides mapping to ENTREZ gene identifiers for Hugo or Ensembl
gene identifiers using either org.Mm.eg.db or org.Mm.eg.db Bioconductor
annotation packages.
}
