% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_genes.R
\name{map_genes}
\alias{map_genes}
\title{Map gene identifiers to ENTREZ gene}
\usage{
map_genes(dat, id = "symbol", species = "human")
}
\arguments{
\item{dat}{Data frame with two columns: \strong{id} and \strong{set_label}.}

\item{id}{A character string specifying the type of gene identifier in the
data, must of be one of two "symbol" (default) or "ensembl".}

\item{species}{A character string specifying the species, must be one of two
"human" (default) or "mouse".}
}
\value{
Data frame with two columns: \strong{entrez} and \strong{set_label}.
The column with the original identifiers is removed.
}
\description{
\code{map_genes} provides mapping to ENTREZ gene identifiers for Hugo or
Ensembl gene identifiers using either \href{https://www.bioconductor.org/packages/release/data/annotation/html/org.Mm.eg.db.html}{org.Mm.eg.db} or \href{https://www.bioconductor.org/packages/release/data/annotation/html/org.Hs.eg.db.html}{org.Hs.eg.db} Bioconductor annotation
packages. Function \code{\link[annotate-getSYMBOL]{annotate::lookUp()}}
extract the mappings. We use org.Xx.eg.db::org.Xx.egALIAS2EG for the mappings
between the common gene symbol idenditifers and
org.Xx.eg.db::org.Xx.egENSEMBL for the mappings between ENTREZ gene
identifiers and Ensembl gene accession numbers.
}
\examples{

\dontrun{
# load test data:
data("human_symbol")
# remove duplicate rows:
dat_clean <- deduplicate_rows(human_symbol)
# map symbol to ENTREZ gene id:
dat_mapped <- map_genes(dat_clean, id = "symbol", species = "human")
}

}
