% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{read_file}
\alias{read_file}
\title{Read data with genes and gene set labels to a data frame}
\usage{
read_file(path, col_names, delim)
}
\arguments{
\item{path}{The name of the fle which the data are to be read from.}

\item{col_names}{A logical value indicating whether the file contains the names
of the variables as its first line.}

\item{delim}{The file field separator character.}
}
\value{
A data frame with two columns, where the first column is 'id' (gene
identifiers) and the second column is 'set.label' (gene set labels).
}
\description{
\code{read_file} reads a file provided a path, column names and file field
delimiter and returns a data frame. For the file to be processed correctly
ensure that the gene identifiers are in the first columns and the set labels
are in the second column in the file.
}
