% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ontologies.R
\name{run_parallel_go}
\alias{run_parallel_go}
\title{Run GO analysis on parallel}
\usage{
run_parallel_go(dat, species = c("human", "mouse"), universe, ontology,
  cores)
}
\arguments{
\item{dat}{data frame with two columns: entrez and set_label. The column
"entrez" should contain Entrez gene identifiers; the column "set_label"
should contain labels for genes that will be analyzed as a group.}

\item{species}{a character string specifying the species, must be one of two "human"
(default) or "mouse". For GO analysis we use 'org.Hs.eg.db' for
human and 'org.Mm.eg.db' for mouse}

\item{ontology}{vector of ontologies for which to run the analysis. If not provided
then the analysis will be run for all ontologies: cellular component (CC),
biological process (BP) and molecular function (MF).}

\item{cores}{numeric value representing the number of cores to use.}
}
\description{
This is a parallel implementation of the function to run GO analysis. The
function will split the supplied data into pieces based on the unique values
in the column "set_label" in the data and will send them to the "cores" for
calculations.
}
